platform: add network namespace support to platform
Add new NMPNetns class. This allows creation, deletion and
switching of network namespaces.

A NMPlatform instance lives inside a namespace, as such, it
has a reference to the NMPNetns instance.

Platform not only uses the netlink socket, but also sysfs, udev,
ethtool, mii. To properly support network namespaces, we must
switch the namespace as necessary. In case of udev, it is only
supported on the main namespace.

For now, network namespaces are not actually used and are disabled
via NM_PLATFORM_NETNS_SUPPORT argument.